+++ frontend/src/models/dto/fullLoanOnboardingRequest.ts
@@ -1,0 +1,7 @@
+import { Borrower } from "./borrower";
+import { createLoanRequest } from "./createLoanRequest";
+
+export interface FullLoanOnboardingRequest {
+  borrower: Borrower;
+  loan: createLoanRequest;
+}