+++ frontend/src/routes/loanConfig/components/CreateLoanModal.tsx
@@ -1,10 +1,11 @@
 import { useState } from "react";
 import FormFactory from "@/components/formComponent/formFactory";
 import { MainApiService } from "@/services/mainApi/mainService";
-import { LoanSummary } from "@/models/dto/loanSummary";
-import { useLoansDispatch, useLoansState } from "@/stores/loans/LoansStore";
+import { useLoansState } from "@/stores/loans/LoansStore";
 import { getTodayDate } from "@/utils/functions/dataTime";
 import { borrowerData } from "@/models/forms/borrowerData";
+import LoadingIndicator from "@/components/loaderComponent/LoaderComponent";
+import { formatName } from "@/utils/functions/strings";
 
 const SuccessView = () => (
    <div className="u-center-v">
@@ -23,13 +24,40 @@ const SaveLoanModal = () => {
   const [formVersion] = useState(0);
   const [loansState, loansDispatch] = useLoansState();
 
-  const createNewLoan = (data: any) => {
-    console.log("CREATE NEW LOAN", loansState, data);
-    // create borrower use borrower model dto
-    // get borrower Id from response
-    // create loan with loan model dto (Remember, status must be pending)
-    // get loan id from response and create schedule with createLoanSchedule model dto
-    // sync loans overview
+  const createNewLoan = (payload: any) => {
+    console.log("CREATE NEW LOAN", loansState, payload);
+    const { simulation: { data }} = loansState;
+    const today = getTodayDate();
+    setStatus('loading');
+    MainApiService.onboardingFullLoan({
+      borrower: {
+        name: formatName(payload.borrowerName),
+        email: payload.email,
+        gender: payload.gender,
+        orgName: payload.organization,
+      },
+      loan: {
+        principal: Number(data.principal),
+        interest_rate: Number(data.interest_rate),
+        term_months: Number(data.term_months),
+        start_date: today,
+        status: 'pending',
+        borrower_id: "",
+        amortization_type: data.amortization_type,
+        payment_frequency: data.payment_frequency
+      },
+    })
+    .then(() => {
+      loansDispatch({
+        type: "SYNC_LOANS_OVERVIEW",
+        isLoaded: false,
+      });
+      setStatus('success');
+    })
+    .catch((error) => {
+      console.error("ERROR CREATING LOAN:", error);
+      setStatus('error');
+    });
 
   }
 
@@ -38,13 +66,16 @@ const SaveLoanModal = () => {
   return(
     <div className="u-center-v">
       <h1 className="paragraph paragraph--lg">Save Loan</h1>
-      <FormFactory
-        key={formVersion}
-        base={borrowerData}
-        formMethod={createNewLoan} 
-        rootCss="form-loan-config"
-        submitLabel="Save"
-      />
+      {status === 'loading' && <LoadingIndicator />}
+      {status === 'idle' &&
+        <FormFactory
+          key={formVersion}
+          base={borrowerData}
+          formMethod={createNewLoan} 
+          rootCss="form-loan-config"
+          submitLabel="Save"
+        />
+      }
     </div>
   )
 }