+++ frontend/src/utils/functions/strings.ts
@@ -1,0 +1,4 @@
+export const formatName = (name: string) => {
+    if(!name) return '';
+    return name.toLowerCase().split(' ').join('_').replace('ñ', 'n');
+}